refactor: transfer types to dedicated module in types directory
